end of ssr for jss

+++ client-jss/src/app/main/services/audio.service.ts
@@ -1,5 +1,6 @@
 import { Injectable } from '@angular/core';
 import { BehaviorSubject } from 'rxjs';
+import { PlatformService } from '../../services/platform.service';
 import { Post } from '../model/Post';
 import { PostService } from './post.service';
 import { UserPreferenceService } from './user.preference.service';
@@ -9,7 +10,7 @@ import { UserPreferenceService } from './user.preference.service';
 })
 export class AudioService {
 
-  private audio = new Audio();
+  private audio: HTMLAudioElement | null = null;
   public isPlaying: boolean = false;
   public currentTime: number = 0;
   public duration: number = 0;
@@ -21,35 +22,39 @@ export class AudioService {
   currentPodcast: Post | undefined;
 
   constructor(private postService: PostService,
-    private userPreferenceService: UserPreferenceService
+    private userPreferenceService: UserPreferenceService,
+    private platformService: PlatformService
   ) {
-    if (this.userPreferenceService.getCurrentPlayingTrackVolume()) {
-      this.audio.volume = parseInt(this.userPreferenceService.getCurrentPlayingTrackVolume()!);
-      this.volume = parseInt(this.userPreferenceService.getCurrentPlayingTrackVolume()!);
-    } else {
-      this.audio.volume = this.volume;
-      userPreferenceService.setCurrentPlayingTrackVolume(this.volume);
-    }
+    this.audio = this.platformService.getNativeAudio();
+    if (this.audio) {
+      if (this.userPreferenceService.getCurrentPlayingTrackVolume()) {
+        this.audio.volume = parseInt(this.userPreferenceService.getCurrentPlayingTrackVolume()!);
+        this.volume = parseInt(this.userPreferenceService.getCurrentPlayingTrackVolume()!);
+      } else {
+        this.audio.volume = this.volume;
+        userPreferenceService.setCurrentPlayingTrackVolume(this.volume);
+      }
 
-    // If already connected, we look in local storage to make it play where the user left it
-    if (userPreferenceService.getCurrentPlayingTrack() && userPreferenceService.getCurrentPlayingTrackTime()) {
-      this.postService.getPostById(parseInt(userPreferenceService.getCurrentPlayingTrack()!)).subscribe(res => {
-        this.setTrackProperties(res);
-        this.audio.currentTime = parseInt(userPreferenceService.getCurrentPlayingTrackTime()!);
-        this.currentTime = this.audio.currentTime;
+      // If already connected, we look in local storage to make it play where the user left it
+      if (userPreferenceService.getCurrentPlayingTrack() && userPreferenceService.getCurrentPlayingTrackTime()) {
+        this.postService.getPostById(parseInt(userPreferenceService.getCurrentPlayingTrack()!)).subscribe(res => {
+          this.setTrackProperties(res);
+          this.audio!.currentTime = parseInt(userPreferenceService.getCurrentPlayingTrackTime()!);
+          this.currentTime = this.audio!.currentTime;
+        });
+      }
+
+      this.audio.addEventListener('timeupdate', () => {
+        this.currentTime = this.audio!.currentTime;
+        userPreferenceService.setCurrentPlayingTrackTime(this.currentTime - 5);
+        this.duration = this.audio!.duration;
+        this.progress = (this.audio!.currentTime / this.audio!.duration) * 100;
+      });
+
+      this.audio.addEventListener('ended', () => {
+        this.isPlaying = false;
       });
     }
-
-    this.audio.addEventListener('timeupdate', () => {
-      this.currentTime = this.audio.currentTime;
-      userPreferenceService.setCurrentPlayingTrackTime(this.currentTime - 5);
-      this.duration = this.audio.duration;
-      this.progress = (this.audio.currentTime / this.audio.duration) * 100;
-    });
-
-    this.audio.addEventListener('ended', () => {
-      this.isPlaying = false;
-    });
   }
 
   // To know if the specific podcast is been played
@@ -68,19 +73,23 @@ export class AudioService {
   }
 
   private setTrackProperties(res: Post) {
-    this.currentPodcast = res;
-    this.isCurrentPodcastDisplayed.next(true);
-    this.audio.src = res.podcastUrl;
-    this.audio.load();
-    this.userPreferenceService.setCurrentPlayingTrack(res.id);
+    if (this.audio) {
+      this.currentPodcast = res;
+      this.isCurrentPodcastDisplayed.next(true);
+      this.audio.src = res.podcastUrl;
+      this.audio.load();
+      this.userPreferenceService.setCurrentPlayingTrack(res.id);
+    }
   }
 
   unloadTrackAndClose() {
-    this.audio.pause();
-    this.currentPodcast = undefined;
-    this.isCurrentPodcastDisplayed.next(false);
-    this.audio.src = "";
-    this.userPreferenceService.deleteAudioPreferences();
+    if (this.audio) {
+      this.audio.pause();
+      this.currentPodcast = undefined;
+      this.isCurrentPodcastDisplayed.next(false);
+      this.audio.src = "";
+      this.userPreferenceService.deleteAudioPreferences();
+    }
   }
 
   togglePlayPause() {
@@ -88,32 +97,37 @@ export class AudioService {
   }
 
   play() {
-    this.audio.play();
+    if (this.audio)
+      this.audio.play();
     this.isPlaying = true;
   }
 
   pause() {
-    this.audio.pause();
+    if (this.audio)
+      this.audio.pause();
     this.isPlaying = false;
   }
 
   // When using the progress bar
   seekTo(value: number) {
-    if (this.audio.duration) {
-      this.audio.currentTime = (value / 100) * this.audio.duration;
-    }
+    if (this.audio)
+      if (this.audio.duration) {
+        this.audio.currentTime = (value / 100) * this.audio.duration;
+      }
   }
 
   // If needs to go at a specific moment of the audio
   addTime(value: number) {
-    if (value) {
-      this.audio.currentTime = this.currentTime + value;
-    }
+    if (this.audio)
+      if (value) {
+        this.audio.currentTime = this.currentTime + value;
+      }
   }
 
   setVolume(value: number) {
     this.volume = value;
-    this.audio.volume = value;
+    if (this.audio)
+      this.audio.volume = value;
     this.userPreferenceService.setCurrentPlayingTrackVolume(value);
   }
 }
